requirements.txt
selenium
beautifulsoup4
requests
lxml
peewee
psycopg2-binary
python-decouple
